Rename controller and model files for GET requests
